[Refactor/#10] SocialType 도메인 이동
