--- app/src/main/java/com/swyp/firsttodo/data/model/type/SocialType.kt
@@ -1,5 +1,0 @@
-package com.swyp.firsttodo.data.model.type
-
-enum class SocialType {
-    GOOGLE,
-}
