+++ app/src/main/java/com/swyp/firsttodo/data/repository/AuthRepository.kt
@@ -1,6 +1,6 @@
 package com.swyp.firsttodo.data.repository
 
-import com.swyp.firsttodo.data.model.type.SocialType
+import com.swyp.firsttodo.domain.model.SocialType
 
 interface AuthRepository {
     suspend fun socialLogin(
