+++ app/src/main/java/com/swyp/firsttodo/domain/model/SocialType.kt
@@ -1,0 +1,5 @@
+package com.swyp.firsttodo.domain.model
+
+enum class SocialType {
+    GOOGLE,
+}
